Update by Feb.10th, 2025
Clear higher 32 bits of CPUID results.
Use generics in page-field calculators.
Fix missing DR6.BS bit in advancing rip in SVM-Core.
There is no need to check debug registers.
For cpuid, use intrinsics instead of inline assembly.
Use a bigger stack for system subverter in Windows.
Add DR7 definitions.

+++ src/svm_core/mod.rs
@@ -449,7 +449,7 @@ impl HypervisorEssentials for SvmHypervisor
 		#[cfg(target_os="uefi")]
 		{
 			let apic_bar=rdmsr(MSR_APIC_BASE);
-			self.mmio_space.add_region(IoRegion::new("lapic",None,svm_apic_output_handler,page_4kb_base(apic_bar as usize) as u64,PAGE_SIZE as u64));
+			self.mmio_space.add_region(IoRegion::new("lapic",None,svm_apic_output_handler,page_4kb_base(apic_bar),PAGE_SIZE as u64));
 		}
 		// Initialize NPT.
 		self.nptm.build_identity_map();
